refactor useGetARmy
